- Fixed missing `type` recipe warning - Simplified jukeboxPlayable to use built in properties method - Added null check for nullable respawn configs in `ServerPlayerKJS#kjs$getSpawnLocation()` - Fixed some InventoryAttachment byte calls - Removed AnimalMixin, calling kjs$foodEaten() from Itemstack#consume instead - Fixed EnergyStorageAttachment/FluidTankAttachment from using legacy addEnergy/removeEnergy/useEnergy which bypassed the transaction system - Fixed another unsafe Optional.get() call in EnergyStorageAttachment - Fixed UtilsJS#wrap where asNumber() and asString() return Optionals and not returning their primitive - Fixed some recursion errors - Fixed `kjs$damageEquipment` unsafe level cast if level is not server - More unsafe Optional.get() calls in 4 different classes - Removed redundant SizedIngredientComponent#FLAT and #OPTIONAL_FLAT since they're now identical to the non-flat variants and are essentially placeholders now - Fixed `KubeRecipe#getFromToString()`` producing malformed output in the form of an extra comma - Fixed InventoryKJS#kjs$clear() method off by one inside index - Deprecated FLAT/OPTIONAL_FLAT with a TODO (not sure if we want to offer support for those in the future) - Added some TODOs

+++ src/main/java/dev/latvian/mods/kubejs/block/entity/InventoryAttachment.java
@@ -126,7 +126,7 @@ public class InventoryAttachment implements BlockEntityAttachment {
 				var tag = ItemStack.CODEC.encodeStart(ops, stack).result().orElse(null);
 
 				if (tag instanceof CompoundTag itemTag) {
-					itemTag.putByte("slot", (byte) i);
+					itemTag.putInt("slot", i);
 					list.add(itemTag);
 				}
 			}
@@ -152,7 +152,7 @@ public class InventoryAttachment implements BlockEntityAttachment {
 				}
 
 				var itemTag = itemTagOpt.get();
-				var slotOpt = itemTag.getByte("slot");
+				var slotOpt = itemTag.getInt("slot");
 				if (slotOpt.isEmpty()) {
 					continue;
 				}
